Use float32 for test_completion.py

+++ tests/v1/entrypoints/openai/test_completion.py
@@ -22,7 +22,7 @@ def default_server_args():
     return [
         # use half precision for speed and memory savings in CI environment
         "--dtype",
-        "bfloat16",
+        "float32",
         "--max-model-len",
         "2048",
         "--max-num-seqs",
